fix(clippy): code style — From/Into, derives, naming, borrows, casts
- Convert impl Into<u64> to impl From for u64 (command_code.rs)
- Derive Default for SMBShareType, add Default impls for
  SMBErrorResponse and SMBIPCShare
- Rename IPC() to ipc() for snake_case convention
- Remove unnecessary borrows, derefs, and same-type casts
- Remove unneeded return statements
- Use iterator instead of index loop in des.rs
- Use Entry API instead of contains_key+insert in server open table

+++ smb/src/protocol/body/filetime.rs
@@ -18,7 +18,7 @@ const TIME_SINCE_1601_AND_EPOCH: u64 = 11644473600000;
 
 impl FileTime {
     pub fn from_unix(unix_timestamp: u64) -> Self {
-        let filetype_normalized = unix_timestamp + TIME_SINCE_1601_AND_EPOCH as u64;
+        let filetype_normalized = unix_timestamp + TIME_SINCE_1601_AND_EPOCH;
         let bytes = u64_to_bytes(filetype_normalized);
         FileTime {
             low_date_time: bytes_to_u32(&bytes[0..4]),
